fix: move testing to isolate
